+++ exporter/jaegerexporter/README.md
@@ -1,9 +1,7 @@
 # Jaeger gRPC Exporter
 
 Exports data via gRPC to [Jaeger](https://www.jaegertracing.io/) destinations.
-By default, this exporter requires TLS and offers queued retry capabilities. A
-Jaeger Thrift HTTP exporter is available in the [contrib
-repository](https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/exporter/jaegerthrifthttpexporter).
+By default, this exporter requires TLS and offers queued retry capabilities.
 
 Supported pipeline types: traces
