+++ exporter/jaegerexporter/go.mod
@@ -1,0 +1,19 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/exporter/jaegerexporter
+
+go 1.16
+
+require (
+	github.com/jaegertracing/jaeger v1.25.0
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal v0.0.0-00010101000000-000000000000
+	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/jaeger v0.0.0-00010101000000-000000000000
+	github.com/stretchr/testify v1.7.0
+	go.opencensus.io v0.23.0
+	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
+	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6
+	go.uber.org/zap v1.19.0
+	google.golang.org/grpc v1.40.0
+)
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal => ../../internal/coreinternal
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/jaeger => ../../pkg/translator/jaeger
