+++ exporter/jaegerexporter/testdata/ca.crt
