+++ exporter/jaegerexporter/testdata/client.crt
