+++ exporter/jaegerexporter/testdata/server.crt
